Support non-string symbolizification, remove !? support
In order to support symbolizification in the same way that most
inflectors seem to work, that is, coercing an object to a string
object, the amount of special cases involved in intelligently
supporting inflected symbolization becomes non-trivial and
non-substainable. Symbolizify will simply convert a string to
symbol-form, and any other additions should be made manually.